Downloads/guild-membership-app-master/guild-membership-app-master/requirements.txt
google-auth
httpclient
six
requests
